fix: 차트 값 소수점 계산 포함
Former-commit-id: 7023e2c2811f3c550f4b7f5edf3bd3a1b086e9d4

+++ client/src/components/BarChart/index.jsx
@@ -4,7 +4,7 @@ import Chart from 'chart.js/auto';
 import { Bar } from 'react-chartjs-2';
 
 function BarChart({ data, width = '100%', height = '100%' }) {
-    const values = data.map((d) => parseInt(d.freq));
+    const values = data.map((d) => parseFloat(d.freq));
     const labels = data.map((d) => d.word);
 
     const backgroundColor = [];
